refactor: rename user controller to auth controller, add user controller, add authenticate middleware
