+++ backend/src/routes/auth.route.ts
@@ -1,0 +1,15 @@
+import { Router } from "express";
+import {
+  createUser,
+  loginUser,
+  logoutUser,
+} from "../controller/auth.controller";
+
+const router = Router();
+
+// routes
+router.post("/register", createUser);
+router.post("/login", loginUser);
+router.get("/logout", logoutUser);
+
+export default router;
