+++ backend/src/routes/index.ts
@@ -1,10 +1,15 @@
 import { Router } from "express";
-import userRoutes from "./user.route";
+import authRoutes from "./auth.route";
 import categoriesRoutes from "./categories.route";
+import userRoutes from "./user.route";
 import { ROUTES } from "./routes";
+import { authenticate } from "../middlewares/authenticate";
 
 const router = Router();
-router.use(ROUTES.AUTH, userRoutes);
+// public routes
+router.use(ROUTES.AUTH, authRoutes);
+// protected routes
 router.use(ROUTES.CATEGORIES, categoriesRoutes);
+router.use(ROUTES.USER, authenticate, userRoutes);
 
 export default router;
