+++ backend/src/routes/routes.ts
@@ -1,8 +1,9 @@
-const BASE_ENDPOINT_PREFIX = "/api";
+const BASE_ENDPOINT_PREFIX = "/api/v1";
 
 const ROUTES = {
   AUTH: `${BASE_ENDPOINT_PREFIX}/auth`,
   CATEGORIES: `${BASE_ENDPOINT_PREFIX}/categories`,
+  USER: `${BASE_ENDPOINT_PREFIX}/user`,
 };
 
 export { ROUTES };
