+++ backend/src/routes/user.route.ts
@@ -1,15 +1,8 @@
 import { Router } from "express";
-import {
-  createUser,
-  loginUser,
-  logoutUser,
-} from "../controller/user.controller";
+import { getUser } from "../controller/user.controller";
 
 const router = Router();
-
 // routes
-router.post("/register", createUser);
-router.post("/login", loginUser);
-router.get("/logout", logoutUser);
+router.get("/getuser", getUser);
 
 export default router;
